# Improve LSM store initialization of shard