--- a/adapters/repos/db/index.go
+++ b/adapters/repos/db/index.go
@@ -449,7 +449,7 @@ func (i *Index) addProperty(ctx context.Context, props ...*models.Property) erro
 	eg.SetLimit(_NUMCPU)
 
 	i.ForEachShard(func(key string, shard ShardLike) error {
-		shard.createPropertyIndex(ctx, eg, props...)
+		shard.initPropertyBuckets(ctx, eg, props...)
 		return nil
 	})