--- a/adapters/repos/db/shard_init_properties.go
+++ b/adapters/repos/db/shard_init_properties.go
@@ -16,47 +16,31 @@ import (
 
 	enterrors "github.com/weaviate/weaviate/entities/errors"
 
-	"github.com/pkg/errors"
 	"github.com/weaviate/weaviate/adapters/repos/db/propertyspecific"
 	"github.com/weaviate/weaviate/entities/models"
 )
 
-func (s *Shard) initProperties(class *models.Class) error {
+func (s *Shard) initProperties(eg *enterrors.ErrorGroupWrapper, class *models.Class) {
 	s.propertyIndices = propertyspecific.Indices{}
 	if class == nil {
-		return nil
+		return
 	}
 
-	eg := enterrors.NewErrorGroupWrapper(s.index.logger)
-	s.createPropertyIndex(context.Background(), eg, class.Properties...)
+	s.initPropertyBuckets(context.Background(), eg, class.Properties...)
 
 	eg.Go(func() error {
-		if err := s.addIDProperty(context.TODO()); err != nil {
-			return errors.Wrap(err, "create id property index")
-		}
-		return nil
+		return s.addIDProperty(context.TODO())
 	})
 
 	if s.index.invertedIndexConfig.IndexTimestamps {
 		eg.Go(func() error {
-			if err := s.addTimestampProperties(context.TODO()); err != nil {
-				return errors.Wrap(err, "create timestamp properties indexes")
-			}
-			return nil
+			return s.addTimestampProperties(context.TODO())
 		})
 	}
 
 	if s.index.Config.TrackVectorDimensions {
 		eg.Go(func() error {
-			if err := s.addDimensionsProperty(context.TODO()); err != nil {
-				return errors.Wrap(err, "crreate dimensions property index")
-			}
-			return nil
+			return s.addDimensionsProperty(context.TODO())
 		})
 	}
-
-	if err := eg.Wait(); err != nil {
-		return errors.Wrapf(err, "init properties on shard '%s'", s.ID())
-	}
-	return nil
 }