## What's being changed:
* Fixes ineffective concurrency
    * Previously, an error group was used, but then new routines were created as part of existing routines, and wait was called multiple times in multiple places
    * Testing this on a test setup with cold page cache could show that there was effectively no concurrency which led to much higher latencies
    * My theory is that if concurrency had worked as intended, other things would have broken. This because there was a loop variable capture (Go gotcha!) that was not caught.
* Fixes inconsistent error annotations
    * annotates errors with the shard just once: previously multiple levels of nesting kept repeating the shard name
* Tries to make the code more readable by
   * splitting up larger methods
   * making at all a bit more consistent
   * splitting up the massive file `shard.go` into separate files for separate lifecycle steps, such as `init`, `drop`, `shutdown`, etc. 

## Speed improvement

The test setup used a real-life dataset with 12 properties and ~1M objects. All OS caches were completely dropped before both runs. The values fluctuated a bit, the below shows one of the best improvements observed:

### Before (`stable/v1.25`)
```
DEBU[0035] loaded non-vector (lsm, object, inverted) in 239.854358ms for shard "testshard0"  action=init_shard_non_vector duration=239.854358ms
```

### After (this branch)
```
DEBU[0019] loaded non-vector (lsm, object, inverted) in 45.581068ms for shard "testshard0"  action=init_shard_non_vector duration=45.581068ms
```
### Review checklist

- [ ] Documentation has been updated, if necessary. Link to changed documentation:
- [ ] Chaos pipeline run or not necessary. Link to pipeline:
- [ ] All new code is covered by tests where it is reasonable.
- [ ] Performance tests have been run or not necessary.